+++ lib/dart_des.dart
@@ -4,7 +4,7 @@ import 'package:dart_des/des_padding.dart';
 enum DESMode { ECB, CBC }
 
 // Types for pad / unpad data
-enum DESPaddingType { None, OneAndZeroes, PKCS7 }
+enum DESPaddingType { None, OneAndZeroes, PKCS7, PKCS5 }
 
 // The base class shared by des and triple des.
 class _BaseDES {
@@ -938,11 +938,12 @@ class DES {
   set iv(List<int> value) => _baseDES.iv = value;
   final DESPaddingType paddingType;
 
-  DES(
-      {required List<int> key,
-      DESMode mode = DESMode.ECB,
-      iv = IV_ZEROS,
-      this.paddingType = DESPaddingType.OneAndZeroes}) {
+  DES({
+    required List<int> key,
+    DESMode mode = DESMode.ECB,
+    iv = IV_ZEROS,
+    this.paddingType = DESPaddingType.OneAndZeroes
+  }) {
     if (key.length != 8) {
       throw Exception(
           'Invalid DES key size. Key must be exactly 8 bytes long.');
@@ -1152,7 +1153,7 @@ class DES3 {
         keySize = 16;
       } else {
         throw Exception(
-            'Invalid triple DES key size. Key must be either 16 or 24 bytes long');
+          'Invalid triple DES key size. Key must be either 16 or 24 bytes long');
       }
     }
     if (mode == DESMode.CBC) {